making cathegories as my second one to many

+++ FortniteSkinWiki2024/app/Http/Controllers/FortniteSkinWikiController.php
@@ -48,6 +48,9 @@ class FortniteSkinWikiController extends Controller
      */
     public function create()
     {
+        if(auth()->user()->admin !== 1){
+            return redirect()->route('FortniteSkinWikis.index')->with('error', 'Access denied');
+        }
         return view('FortniteSkinWikis.create');
     }
 
@@ -94,6 +97,10 @@ class FortniteSkinWikiController extends Controller
      */
     public function edit(FortniteSkinWiki $FortniteSkinWiki)
     {
+        if(auth()->user()->admin !== 1){
+            return redirect()->route('FortniteSkinWikis.index')->with('error', 'Access denied');
+        }
+        
         return view('FortniteSkinWikis.edit', compact('FortniteSkinWiki'));
     }
 
@@ -117,6 +124,9 @@ class FortniteSkinWikiController extends Controller
      */
     public function destroy(FortniteSkinWiki $FortniteSkinWiki)
     {
+        if(auth()->user()->admin !== 1){
+            return redirect()->route('FortniteSkinWikis.index')->with('error', 'Access denied');
+        }
         $FortniteSkinWiki->delete();
     
         
@@ -128,5 +138,10 @@ class FortniteSkinWikiController extends Controller
     public function reviews()
     {
         return $this->hasMany(Review::class);
+    }
+
+    public function cathegories()
+    {
+        return $this->hasMany(Cathegory::class);
     }
 }
